+++ src/components/Results/style.css
