+++ src/components/Search/style.css
